feat: Add endpoint for MTTR last 7 days
Adds a new endpoint `/api/metrics/tails/mttr/last-7-days` to provide Mean Time To Resolve (MTTR) data for TailEntity objects over the past 7 days.

The endpoint returns data in the following format:
{
  "labels": ["MMM dd", ...],
  "data": [mttr_for_day_1_in_hours, ...]
}

Changes include:
- Modified `TailDatasetResponse` to use `List<Double>` for data.
- Added `getTailMTTRLast7Days()` to `TailMetricsService` interface.
- Implemented `getTailMTTRLast7Days()` in `TailMetricsServiceImpl`, including logic to calculate daily MTTR in hours.
- Added the new GET endpoint in `TailMetricsController`.
- Unit tests were skipped as you requested.

+++ n1netails-api/src/main/java/com/n1netails/n1netails/api/model/response/TailDatasetResponse.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class TailDatasetResponse {
     private String label;
-    private List<Integer> data;
+    private List<Double> data;
 }
